0.1.5 Release
-Persistent data now saves at the same time the system is saved. This
was to resolve an issue where the game froze or crashed shortly after
financials were processed but before the game saved.
-Custom intervals! you may now specify your own period in Kerbin days
from 1-9999. All values are calculated accordingly.
-Autopay is now working! You can specify the percentage of how  much of
your available Funds can be paid out automatically.
-Loans are now NOT paid automatically. They are also not figured into
the Autopay system, they must be paid manually. This is to make way for
upcoming changes in 0.2.0.